Backend moved filter elements from subfolders to root and renamed accordingly
