+++ frontend/src/visualization/filtering/helpers/FilterPropertyModel.ts
@@ -1,5 +1,5 @@
-import { Property } from '../../shared/entities';
-import { FilterModelEntry } from '../../shared/filter';
+import { Property } from '../../../shared/entities';
+import { FilterModelEntry } from '../../../shared/filter';
 
 export default interface FilterPropertyModel extends FilterModelEntry {
   selectedValues: Property[] | null;
